Moved .exists(Predicate) and .forAll(Predicate) from Iterable to Traversable
