+++ src/main/java/io/vavr/Iterable.java
@@ -22,7 +22,6 @@ import io.vavr.collection.Iterator;
 
 import java.util.Objects;
 import java.util.function.Function;
-import java.util.function.Predicate;
 
 /**
  * Extension of the well-known Java {@link java.lang.Iterable} in the sense that a rich
@@ -40,35 +39,6 @@ public interface Iterable<T> extends java.lang.Iterable<T> {
      */
     @Override
     Iterator<T> iterator();
-
-    /**
-     * Checks, if an element exists such that the predicate holds.
-     *
-     * @param predicate A Predicate
-     * @return true, if predicate holds for one or more elements, false otherwise
-     * @throws NullPointerException if {@code predicate} is null
-     */
-    default boolean exists(Predicate<? super T> predicate) {
-        Objects.requireNonNull(predicate, "predicate is null");
-        for (T t : this) {
-            if (predicate.test(t)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Checks, if the given predicate holds for all elements.
-     *
-     * @param predicate A Predicate
-     * @return true, if the predicate holds for all elements, false otherwise
-     * @throws NullPointerException if {@code predicate} is null
-     */
-    default boolean forAll(Predicate<? super T> predicate) {
-        Objects.requireNonNull(predicate, "predicate is null");
-        return !exists(predicate.negate());
-    }
 
     // `Iterable<T>` must not have a generic type bound, see TraversableTest ShouldJustCompile
     default <C> C to(Function<? super java.lang.Iterable<T>, C> fromIterable) {
